Move binary operations to their own package
