+++ pkg/streamingpromql/operators/binops/binary_operation.go
@@ -1,6 +1,6 @@
 // SPDX-License-Identifier: AGPL-3.0-only
 
-package operators
+package binops
 
 import (
 	"slices"
